+++ app/Providers/RepositoryServiceProvider.php
@@ -2,7 +2,9 @@
 
 namespace App\Providers;
 
+use App\Repositories\BrasileiraoJogosRepository;
 use App\Repositories\BrasileiraoRepository;
+use App\Repositories\Contracts\BrasileiraoJogosRepositoryInterface;
 use App\Repositories\Contracts\BrasileiraoRepositoryInterface;
 use Illuminate\Support\ServiceProvider;
 
@@ -17,6 +19,9 @@ class RepositoryServiceProvider extends ServiceProvider
     {
         /** Brasileirao Repository */
         $this->app->bind(BrasileiraoRepositoryInterface::class, BrasileiraoRepository::class);
+
+        /** Brasileirao Jogos Repository */
+        $this->app->bind(BrasileiraoJogosRepositoryInterface::class, BrasileiraoJogosRepository::class);
     }
 
     /**